fix(ci): resolve all clippy and docs lint failures
xberg-gliner-candle clippy:
- encoder.rs: #[allow(dead_code)] on hidden_size (pub API surface for future use)
- lora.rs: #[allow(dead_code)] on target_modules (config field preserved for round-trip)
- pipeline.rs: remove unnecessary 'as usize' cast (max_position_embeddings is already usize)
- heads/count_pred.rs, scorer.rs, span_rep.rs: blank line before Returns in doc comments
- lora.rs: collapse nested if-let into let chain (Rust 2024 edition)
- model.rs: collapse nested if/if-let into single let chain

docs lint:
- .textlintignore: exclude docs/superpowers/ (internal planning docs, not user-facing)
- .lychee.toml: add exclude_path for docs/superpowers/ (plan files may reference unbuilt targets)
- 2026-06-29-xberg-mcp-rag-gdpr.md: convert link to code span inside fenced code example

+++ crates/xberg-gliner-candle/src/pipeline.rs
@@ -29,7 +29,7 @@ pub(crate) fn run_pipeline(
     let encoded = encode_v2(text, labels, tokenizer, splitter)?;
 
     // 1. Truncate to the encoder's position-embedding limit.
-    let max_seq = encoder.config.max_position_embeddings as usize;
+    let max_seq = encoder.config.max_position_embeddings;
     let seq_len = encoded.input_ids.len().min(max_seq);
     let input_ids = Tensor::from_slice(&encoded.input_ids[..seq_len], (1, seq_len), device)?;
     let attn_data: Vec<i64> = vec![1_i64; seq_len];
